+++ src/main.rs
@@ -37,7 +37,11 @@ fn rocket() -> _ {
                 user_route::get_user_by_id,
                 user_route::update_user_by_id,
                 user_route::delete_user_by_id,
-                post_route::insert_single_post
+                post_route::insert_single_post,
+                post_route::get_all_posts,
+                post_route::get_post_by_id,
+                post_route::update_post_by_id,
+                post_route::delete_post_by_id
             ],
         )
         .attach(cors)