+++ src/services/impl_service/task_impl.rs
@@ -3,16 +3,24 @@ use crate::mappers::task_mapper;
 use crate::models::task::{NewTask, Task};
 use crate::services::task_service;
 impl task_service::GetTask for Task {
-    fn insert_single_task(task: NewTask) -> Result<Task, diesel::result::Error> {
+    // GOOD:
+    fn insert_single_task(task: &NewTask) -> Result<Task, Box<dyn std::error::Error>> {
         match establish_pg_connection() {
-            Ok(mut conn) => match task_mapper::insert_task(&mut conn, &task) {
+            Ok(mut conn) => match task_mapper::insert_task(&mut conn, task) {
                 Ok(inserted_task) => Ok(inserted_task),
-                Err(e) => Err(e),
+                Err(e) => {
+                    println!("{e:?}");
+                    Err(Box::new(e))
+                }
             },
-            Err(_) => Err(diesel::result::Error::NotFound),
+            Err(e) => {
+                println!("{e:?}");
+                Err(Box::new(e))
+            }
         }
     }
-    fn get_all_tasks() -> Result<Vec<Task>, diesel::result::Error> {
+    // GOOD:
+    fn get_all_tasks() -> Result<Vec<Task>, Box<dyn std::error::Error>> {
         match establish_pg_connection() {
             Ok(mut conn) => {
                 match task_mapper::fetch_all_tasks(&mut conn) {
@@ -20,55 +28,90 @@ impl task_service::GetTask for Task {
                         if all_tasks.len() != 0 {
                             Ok(all_tasks)
                         } else {
-                            Err(diesel::result::Error::NotFound)
+                            println!("Empty query.");
+                            Err(Box::new(diesel::result::Error::NotFound))
                         }
                     }
                     Err(e) => {
                         // panic!("oWo! Please add task first!");
-                        Err(e)
+
+                        println!("{e:?}");
+                        Err(Box::new(e))
                     }
                 }
             }
-            Err(_) => Err(diesel::NotFound),
+            Err(e) => {
+                println!("{e:?}");
+                Err(Box::new(e))
+            }
         }
     }
-    fn get_task_by_id(t_id: i32) -> Result<Task, diesel::result::Error> {
+    // GOOD:
+    fn get_task_by_id(t_id: i32) -> Result<Task, Box<dyn std::error::Error>> {
         match establish_pg_connection() {
             Ok(mut conn) => match task_mapper::fetch_task_by_id(&mut conn, t_id) {
                 Ok(task) => Ok(task),
-                Err(e) => Err(e),
+                Err(e) => {
+                    println!("{e:?}");
+                    Err(Box::new(e))
+                }
             },
-            Err(_) => Err(diesel::NotFound),
+            Err(e) => {
+                println!("{e:?}");
+                Err(Box::new(e))
+            }
         }
     }
+    // GOOD:
     fn update_task_by_id(
         t_id: i32,
-        task: crate::models::task::PatchTask,
-    ) -> Result<Task, diesel::result::Error> {
+        task: &crate::models::task::PatchTask,
+    ) -> Result<Task, Box<dyn std::error::Error>> {
         match establish_pg_connection() {
-            Ok(mut conn) => match task_mapper::update_task_by_id(&mut conn, t_id, task) {
+            Ok(mut conn) => match task_mapper::update_task_by_id(&mut conn, t_id, &task) {
                 Ok(task) => Ok(task),
-                Err(e) => Err(e),
+                Err(e) => {
+                    println!("{e:?}");
+                    Err(Box::new(e))
+                }
             },
-            Err(_) => Err(diesel::NotFound),
+            Err(e) => {
+                println!("{e:?}");
+                Err(Box::new(e))
+            }
         }
     }
-    fn delete_task_by_id(t_id: i32) -> Result<Task, diesel::result::Error> {
+    /// ## Service impl
+    /// GOOD:
+    fn delete_task_by_id(t_id: i32) -> Result<Task, Box<dyn std::error::Error>> {
         match establish_pg_connection() {
             Ok(mut conn) => match task_mapper::delete_task_by_id(&mut conn, t_id) {
                 Ok(deleted_task) => Ok(deleted_task),
-                Err(e) => Err(e),
+                Err(e) => {
+                    println!("{e:?}");
+                    Err(Box::new(e))
+                }
             },
-            Err(_) => Err(diesel::NotFound),
+            Err(e) => {
+                println!("{e:?}");
+                Err(Box::new(e))
+            }
         }
     }
-    fn insert_full_single_task(task: Task) -> Result<Task, diesel::result::Error> {
+    // GOOD:
+    fn insert_full_single_task(task: &Task) -> Result<Task, Box<dyn std::error::Error>> {
         match establish_pg_connection() {
-            Ok(mut conn) => match task_mapper::insert_full_single_task(&mut conn, &task) {
+            Ok(mut conn) => match task_mapper::insert_full_single_task(&mut conn, task) {
                 Ok(inserted_full_task) => Ok(inserted_full_task),
-                Err(e) => Err(e),
+                Err(e) => {
+                    println!("{e:?}");
+                    Err(Box::new(e))
+                }
             },
-            Err(_) => Err(diesel::NotFound),
+            Err(e) => {
+                println!("{e:?}");
+                Err(Box::new(e))
+            }
         }
     }
 }
@@ -87,8 +130,9 @@ mod tests {
             None,
             Some(get_e8_time()),
             Some(get_e8_time()),
+            Some(4),
         );
-        let _ = Task::insert_single_task(task);
+        let _ = Task::insert_single_task(&task);
     }
 
     #[test]
@@ -110,8 +154,9 @@ mod tests {
             "new title for put task".to_string(),
             "hello".to_string().into(),
             Some(get_e8_time()),
+            Some(4),
         );
-        let updated_task = Task::update_task_by_id(t_id, task);
+        let updated_task = Task::update_task_by_id(t_id, &task);
         println!("updated_task: {updated_task:?}");
     }
     #[test]
@@ -121,8 +166,13 @@ mod tests {
     }
     #[test]
     fn test_insert_full_single_task() {
-        let task = Task::new(2, "title1".to_string(), "content".to_string().into());
-        let inserted_task = Task::insert_full_single_task(task);
+        let task = Task::new(
+            2,
+            "title1".to_string(),
+            "content".to_string().into(),
+            Some(4),
+        );
+        let inserted_task = Task::insert_full_single_task(&task);
         println!("{inserted_task:?}");
     }
 }