+++ src/services/impl_service/user_impl.rs
@@ -8,50 +8,80 @@ use crate::{
 impl user_service::GetUser for User {
     fn insert_single_user(
         user: &crate::models::user::NewUser,
-    ) -> Result<User, diesel::result::Error> {
+    ) -> Result<User, Box<dyn std::error::Error>> {
         match establish_pg_connection() {
             Ok(mut conn) => match user_mapper::insert_user(&mut conn, user) {
                 Ok(user) => Ok(user),
-                Err(e) => Err(e),
+                Err(e) => {
+                    println!("{e:?}");
+                    Err(Box::new(e))
+                }
             },
-            Err(_) => Err(diesel::NotFound),
+            Err(e) => {
+                println!("{e:?}");
+                Err(Box::new(e))
+            }
         }
     }
-    fn get_all_users() -> Result<Vec<User>, diesel::result::Error> {
+    fn get_all_users() -> Result<Vec<User>, Box<dyn std::error::Error>> {
         match establish_pg_connection() {
             Ok(mut conn) => match user_mapper::fetch_all_users(&mut conn) {
                 Ok(all_users) => Ok(all_users),
-                Err(e) => Err(e),
+                Err(e) => {
+                    println!("{e:?}");
+                    Err(Box::new(e))
+                }
             },
-            Err(_) => Err(diesel::NotFound),
+            Err(e) => {
+                println!("{e:?}");
+                Err(Box::new(e))
+            }
         }
     }
 
-    fn get_user_by_id(id: i32) -> Result<User, diesel::result::Error> {
+    fn get_user_by_id(id: i32) -> Result<User, Box<dyn std::error::Error>> {
         match establish_pg_connection() {
             Ok(mut conn) => match user_mapper::fetch_user_by_id(&mut conn, id) {
                 Ok(user) => Ok(user),
-                Err(e) => Err(e),
+                Err(e) => {
+                    println!("{e:?}");
+                    Err(Box::new(e))
+                }
             },
-            Err(_) => Err(diesel::NotFound),
+            Err(e) => {
+                println!("{e:?}");
+                Err(Box::new(e))
+            }
         }
     }
-    fn update_user_by_id(id: i32, user: &PatchUser) -> Result<User, diesel::result::Error> {
+    fn update_user_by_id(id: i32, user: &PatchUser) -> Result<User, Box<dyn std::error::Error>> {
         match establish_pg_connection() {
             Ok(mut conn) => match user_mapper::update_user_by_id(&mut conn, id, user) {
                 Ok(updated_user) => Ok(updated_user),
-                Err(e) => Err(e),
+                Err(e) => {
+                    println!("{e:?}");
+                    Err(Box::new(e))
+                }
             },
-            Err(_) => Err(diesel::NotFound),
+            Err(e) => {
+                println!("{e:?}");
+                Err(Box::new(e))
+            }
         }
     }
-    fn delete_user_by_id(id: i32) -> Result<User, diesel::result::Error> {
+    fn delete_user_by_id(id: i32) -> Result<User, Box<dyn std::error::Error>> {
         match establish_pg_connection() {
             Ok(mut conn) => match user_mapper::delete_user_by_id(&mut conn, id) {
                 Ok(deleted_user) => Ok(deleted_user),
-                Err(e) => Err(e),
+                Err(e) => {
+                    println!("{e:?}");
+                    Err(Box::new(e))
+                }
             },
-            Err(_) => Err(diesel::NotFound),
+            Err(e) => {
+                println!("{e:?}");
+                Err(Box::new(e))
+            }
         }
     }
 }